Subhoshrighoshtask4.md
Here we will be adding our day 1,  day 2  and day 3 task

##  TASK 1

Answers to the Questions given :-


1. What are Git and GitHub?
 Ans - GIT :
☆☆ Git is a version control system that lets us to manage and keep track of our source code history.

☆☆ one thing that really sets Git apart is its branching model. Branching allows us to create independent local branches in our code. This means we can try out new ideas, set aside branches for production work, jump back to earlier branches, and easily delete, merge, and recall branches at the click of a button.

GITHUB:
☆☆ GitHub is a cloud based hosting service that let us manage Git Repositories.

☆☆ It presents an extremely intuitive, graphically represented user interface, and provides programmers with built-in control and task-management tools.


2. Why GitHub is so popular and used in most of the projects?

Ans - It is so because Through GitHub, we can share our code with others, giving them the power to make revisions or edits on our various Git branches. This makes it possible for entire teams to coordinate together on single projects in real-time. As changes are introduced, new branches are created, allowing the team to continue to revise the code without overwriting each other’s work. These branches are like copies, and changes made on them do not reflect in the main directories on other users’ machines unless users choose to push/pull the changes to incorporate them. There is also a GitHub desktop application available, which offers some additional functionality for experienced developers.This is why GitHub is so popular and used in most of the projects.

3. What are the other platforms similar to GitHub?

Ans - 
☆ GitLab
☆ BigBucket
☆ Source Forge

##  TASK 2


(1). How a git workflow works ?

Ans - Workflow has 4 fundamentals -
A. Workspace
B. Index
C. local repository 
D. remote repository

A. Workspace: The Working Directory
B. Index : A staging area between workspace and the repository
C. Local Repository : Here we make the local changes
D. Remote Repository : It is on the Server.


(2). What are different stages of a git project as  commit,add?

Ans - A. Modified : The changes can be made in the code without further changing the actual code.
B. Staged: staging area or index, the place where all the new files are finally ready to be joined to the remote repository.
C. Committed : It means the current changes in the data is modified and safely stored in local repository.

(3). Is it possible to do a git commit before git add. If you have made any changes? Explain why?

Ans - No It's not possible to do a git commit before git add.
As files are added to the index to make commits. According to the git workflow _ At first, files are added  to the index and then it is committed.

(4). Why is git diff used ?

Ans - git diff command is used to check or track the difference between the files , where they differ i.e., in the stage and the workspace.

(5). Can we leave the commit message as blank ?

Ans - Yes we can leave the commit message as blank .

 By using the command :
git commit -a --allow-empty -message -m ""

My github repo link is :
 https://github.com/subhoshrighosh/git_learning


## TASK 3



A. What is meant by the term fork and clone?

Ans ---》 FORK :- A fork is a copy of a repository. Forking a repository allows you to freely experiment with changes without affecting the original project.

Most commonly, forks are used to either propose changes to someone else's project or to use someone else's project as a starting point for your own idea.

CLONE :- Clone is same as what the word has in the literary terms, making a copy of the files present in the cloud on github server to your local machine.
Cloning basically means you want to get a local copy of the code present in the repository. After cloning you can then do whatever changes you like in the code and then you can pull the changes back to the repository.

** fork is in essence a clone too but it doesn't keep up with the original repo. So you will not be able to track the changes made to the original repo when you do a 'git fetch', 'git status' and finally 'git merge' (or 'git pull') .

B. What are branches in github ?

 Ans - Git branch is a feature in git used for separating a feature or part of code from you your master in order not to mess something in your main code like suppose you have a website as a project in git and you want to add a new feature to it so you do it by creating a new branch for it so your main code remain the same as it and once you complete it then you merge it with your master.

C. What is PR ?

Ans - PR stands for pull request. A pull/merge request is submitted when you’ve worked on some code from a particular branch and want to inform the others of the changes you’ve made. A person/people can be assigned to review and subsequently approve the request before your changes can be incorporated into the branch. Do note that it is referred to as a “pull-request” on GitHub and Bitbucket, but other platforms like GitLab and Gitorious have opted to call it a “merge-request”.

D. Can we delete the master branch if not why?

Ans - yes we can , but only after creating a new branch and set that branch as the default one.


E. How can we delete a branch ? 

Ans - *DELETING A BRANCH*

## Delete a remote branch

a. git push origin --delete <branch> 
 ( Git version 1.7.0 or newer )

b. git push origin :<branch> 

(Git versions older than 1.7.0)


## Delete a local branch

a. git branch --delete <branch>

b. git branch -d <branch>
 
( Shorter version)

C. git branch -D <branch>

 (Force delete un-merged branches)

{☆Also in github, we can have head branches automatically deleted after pull requests are merged in your repositor.}

*LINKS*

https://github.com/deepak2431/gitseries/pull/10/commits


https://github.com/deepak2431/gitseries/pull/12


Regards
*SUBHOSHRI GHOSH*


    -[Task 4 ends here]